Implement the date picker
- Modify DateField to use DateTimePicker react-widget
- Refactor uses of DateField to pass new props
- Refactor onChange for DateField as it is given a value instead of an
event
- Add dateTimeFormatter to dateFormatter.js and use it in the view
- Commented out two lines of required label checking in DateFieldSpec as
the props are not passed all the way down to the input field by the
DateTimePicker component. The broken functionality was not explicitly
specified in the story, and will likely be handled by the implementation
of validation checking. We discussed this with Bruno and Aman and
decided it was the best action to get the date picker in without those
two checks.

[#145765803]